requirements/base.txt
Django==1.6
South==0.8
django-auth-ldap==1.1.5
django-floppyforms==1.1
django-floppy-gumby==0.1
requests==2.0.1
beautifulsoup4==4.3.2
Pillow==2.2.1
djangorestframework==2.3.8
django-filter==0.7